Makefile: fix webui target
The webui target is not a file. This causes the webui
target and all targets that depends on it to be always
rebuilded, which is not desirable.

Most notably, this causes a second compilation at
'make install' when using the traditional approach
to build software with 'make && make install', which
should not occur.

This commit fixes this situation and now 'make install'
only install files if they were already built.

Signed-off-by: Daniel Bermond <danielbermond@gmail.com>